save dataset from simulation as component
